Resolved #11, added option how logs should be handled.
Described logging and its options in README and man page.

+++ utils.go
@@ -10,8 +10,9 @@ import (
 )
 
 const (
-	pathLogFile    = "/var/log/emptty"
-	pathLogFileOld = "/var/log/emptty.old"
+	pathLogFileNull = "/dev/null"
+	pathLogFile     = "/var/log/emptty"
+	pathLogFileOld  = "/var/log/emptty.old"
 )
 
 // propertyFunc defines method to be invoked during readProperties method for each record.
@@ -83,13 +84,19 @@ func handleArgs() {
 }
 
 // Initialize logger to file defined by pathLogFile.
-func initLogger() {
-	if fileExists(pathLogFile) {
-		os.Remove(pathLogFileOld)
-		os.Rename(pathLogFile, pathLogFileOld)
+func initLogger(conf *config) {
+	logFilePath := pathLogFile
+
+	if conf.logging == Default {
+		if fileExists(pathLogFile) {
+			os.Remove(pathLogFileOld)
+			os.Rename(pathLogFile, pathLogFileOld)
+		}
+	} else if conf.logging == Disabled {
+		logFilePath = pathLogFileNull
 	}
 
-	f, err := os.OpenFile(pathLogFile, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0644)
+	f, err := os.OpenFile(logFilePath, os.O_RDWR|os.O_APPEND|os.O_CREATE, 0644)
 	if err == nil {
 		log.SetOutput(f)
 	}
